Added bullets to alien_invasion.py & game_functions.py
Stored bullets in a Group in alien_invasion.py to draw bullets to the
screen on each pass through the main loop and to update each bullet's
position.

In game_functions.py, modified check_keydown_events() to fire a bullet
when the space bar is pressed. We don't need to update
check_keyup_events() because nothing happens when the key is released.
Modified update_screen() to make sure each bullet is redrawn to the
screen before we call flip().