Instantiated and drew the Alien class on-screen
We will continue to use bitmap images for simplicity. The creation of
the alien class is similar to the ship class. The major difference being
the starting position of the alien craft, i.e. top-left corner of the
screen.

The alien instance was also created in alien_invasion.py and the craft's
blitme() was added to game_functions.py.