Moving the fleet to the right
Created an update() method in alien.py and an alien_speed_factor
in settings.py.

The self.x attribute, which can hold decimal values, tracks the fleet's
position.

In the main while loop, added a call to update the position of aliens.

Added update_aliens() at the end of game_functions