Added a game_active flag to game_stats
Added a game_active flag to game_stats.py.

Changed ship_hit() in game_functions.py so that game_active changes to
False if the player has used up all their ships.

Changed alien_invasion.py by adding an if block in the main loop which
detects whether the game_active flag is true or not.